WIP radiasoft/download#350: configure git pull to not check safe directory
Fedora 36 has git 2.39.2 which checks to make sure that the user running
the git operations is the owner of the git directory (more explanation https://github.com/git/git/commit/8959555cee7ec045958f9b6dd62e541affb7e7d9).
This causes problems when working over an NFS mount because from git's
perspective the user on the NFS client may not be the owner of the directory.
The git pull here is the only git operation we do in our install that is
impacted so allow configuration to be supplied to turn off the check.